Add constraint checks

+++ CRM/Civirules/Upgrader.php
@@ -37,7 +37,7 @@ class CRM_Civirules_Upgrader extends CRM_Civirules_Upgrader_Base {
       }
     }
 
-    if (CRM_Core_DAO::checkTableExists("civicrm_event")) {
+    if (CRM_Core_DAO::checkTableExists("civirule_event")) {
       CRM_Core_DAO::executeQuery("
         INSERT INTO civirule_event (name, label, object_name, op, cron, class_name, created_date, created_user_id)
         VALUES
@@ -64,16 +64,25 @@ class CRM_Civirules_Upgrader extends CRM_Civirules_Upgrader_Base {
     }
     // rename columns event_id and event_params in civirule_rule
     if (CRM_Core_DAO::checkTableExists("civirule_rule")) {
-      CRM_Core_DAO::executeQuery("ALTER TABLE civirule_rule DROP FOREIGN KEY fk_rule_event;");
-      CRM_Core_DAO::executeQuery("ALTER TABLE civirule_rule DROP INDEX fk_rule_event_idx;");
+      $this->ctx->log->info('civirules 1002: Drop fk_rule_event, fk_rule_event_idx.');
+      if (CRM_Core_DAO::checkConstraintExists('civirule_rule', 'fk_rule_event')) {
+        CRM_Core_DAO::executeQuery("ALTER TABLE civirule_rule DROP FOREIGN KEY fk_rule_event;");
+      }
+      if (CRM_Core_DAO::checkConstraintExists('civirule_rule', 'fk_rule_event_idx')) {
+        CRM_Core_DAO::executeQuery("ALTER TABLE civirule_rule DROP INDEX fk_rule_event_idx;");
+      }
       if (CRM_Core_DAO::checkFieldExists('civirule_rule', 'event_id')) {
         CRM_Core_DAO::executeQuery("ALTER TABLE civirule_rule CHANGE event_id trigger_id INT UNSIGNED;");
       }
       if (CRM_Core_DAO::checkFieldExists('civirule_rule', 'event_params')) {
         CRM_Core_DAO::executeQuery("ALTER TABLE civirule_rule CHANGE event_params trigger_params TEXT;");
       }
-      CRM_Core_DAO::executeQuery("ALTER TABLE civirule_rule ADD CONSTRAINT fk_rule_trigger FOREIGN KEY (trigger_id) REFERENCES civirule_trigger(id);");
-      CRM_Core_DAO::executeQuery("ALTER TABLE civirule_rule ADD INDEX fk_rule_trigger_idx (trigger_id);");
+      if (!CRM_Core_DAO::checkConstraintExists('civirule_rule', 'fk_rule_trigger')) {
+        CRM_Core_DAO::executeQuery("ALTER TABLE civirule_rule ADD CONSTRAINT fk_rule_trigger FOREIGN KEY (trigger_id) REFERENCES civirule_trigger(id);");
+      }
+      if (!CRM_Core_DAO::checkConstraintExists('civirule_rule', 'fk_rule_trigger_idx')) {
+        CRM_Core_DAO::executeQuery("ALTER TABLE civirule_rule ADD INDEX fk_rule_trigger_idx (trigger_id);");
+      }
     }
     return true;
   }
